Create decisions.md to document permissions
It was recently decided that we should give VFS teams write access to this repo. We still expect them to follow content review procedures with VSP, but are not (currently) enforcing that via a technical block at the point of code review. This file documents that decision and who endorsed it.